Убрал тикеты
