--- backend/src/domain/usecases/tickets.py
@@ -1,24 +1,0 @@
-from abc import ABC, abstractmethod
-
-from schemas.payment import RefundSchema
-
-
-class AbstractTicketUseCase(ABC):
-
-    @abstractmethod
-    async def get_info(self, ticket_id: int, user_id: int): ...
-
-    @abstractmethod
-    async def get_list_by_current_user(self, user_id: int): ...
-
-    @abstractmethod
-    async def get_list_by_user(self, target_user_id: int, user_id: int): ...
-
-    @abstractmethod
-    async def get_list(self, user_id: int): ...
-
-    @abstractmethod
-    async def refund(self, ticket_id: int, user_id: int, req: RefundSchema): ...
-
-    @abstractmethod
-    async def delete(self, ticket_id: int, user_id: int): ...
