+++ backend/src/domain/utils/unit_of_work.py
@@ -2,7 +2,6 @@ from abc import ABC, abstractmethod
 
 from repositories.bookings import BookingsRepo
 from repositories.events import EventsRepo
-from repositories.tickets import TicketsRepo
 from repositories.users import UsersRepo
 
 
@@ -10,7 +9,6 @@ class AbstractUOW(ABC):
     users: UsersRepo
     events: EventsRepo
     bookings: BookingsRepo
-    tickets: TicketsRepo
 
     @abstractmethod
     async def __aenter__(self): ...
