--- backend/src/repositories/tickets.py
@@ -1,8 +1,0 @@
-from sqlalchemy import delete
-
-from models.tickets import Tickets
-from repositories.alchemy import SqlAlchemyRepo
-
-
-class TicketsRepo(SqlAlchemyRepo):
-    model = Tickets
