--- backend/src/schemas/tickets.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-
-
-class TicketSchema(BaseModel):
-    id: int
-    booking_id: int
