+++ backend/src/utils/unit_of_work.py
@@ -1,7 +1,6 @@
 from db.database import async_session_maker
 from domain.utils.unit_of_work import AbstractUOW
 from repositories.bookings import BookingsRepo
-from repositories.tickets import TicketsRepo
 from repositories.users import UsersRepo
 from repositories.events import EventsRepo
 
@@ -12,7 +11,6 @@ class UOW(AbstractUOW):
         self.users = UsersRepo(self.session)
         self.events = EventsRepo(self.session)
         self.bookings = BookingsRepo(self.session)
-        self.tickets = TicketsRepo(self.session)
 
     async def __aexit__(self, *args):
         await self.rollback()
